Add files via upload
Fully implemented my database as well as fully implemented my GUI. Next step I will add another thing to be tracked by the database, it will be very simple but important for the functionality of the app. After that I plan to make the app downloadable by localizing the database into one file as well as making it executable.